requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
google-generativeai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
langserve
langchain_community
uvicorn
pymupdf
langchain_groq
pytube
numexpr
langchain
chromadb
groq